Split types in separate modules.
